Calculating Easter
These programs are designed to take an input year from user which then prints out the exact month and day of Easter in that year.  The Node.java and LinkedList.java are for calculating the number of times each month and day repeated looping through a number of years.  In this case, the years are from 0-5,700,000.